LayoutTests/imported/w3c/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/015-expected.txt


PASS global scope unchanged
PASS window object unchanged
PASS this is the window object
